!fixup Move Storage into its own package
