+++ src/main/java/dev/jingyen/duke/Duke.java
@@ -4,6 +4,7 @@ import dev.jingyen.duke.model.Task;
 import dev.jingyen.duke.parser.InvalidInputException;
 import dev.jingyen.duke.parser.InvalidTaskException;
 import dev.jingyen.duke.parser.TaskParser;
+import dev.jingyen.duke.storage.Storage;
 
 import java.io.File;
 import java.io.IOException;
